Added endpoint to run all data definitions and changed some expected values for tests.
The new endpoint is 'get-all/typical'. When running the unit tests I saw that most of the results
were off by <100, which was likely due to David adding new data to the source database.
Since the scope of the queries was so large and the results were only off by a small amount, I updated the expected values of
the test results to the new numbers I was getting.